Add impl namespace
Moves most of the implementation into a private namespace, this allows
creating a more focused core API which can focus on assembling the
pieces provided by the impl namespace.

Over time, I expect parts of the impl to move into a documented
namespace of some kind to allow for alternative assemblies of systems.