Remove xforms based implementation
This suffered from a few drawbacks:

*Hard to implement async between xforms*

High would like to make async transparent between multiple executions on
a single component.  pre-start -> start -> post-start should be
transparent to the authors of these functions about the async execution
that is taking place.  The use of a vector to create a faux xform was a
workaround that didn't fit the spirit of xforms.

*Used a non-standard rf form*

The other way in which high xforms were not real xforms is that they
used a customized rf format.  This was to support the various additional
metadata an xform wanted in order to run.

*Complicated implementation*

The implementation of promesa-run was complicated and hard to
understand.  This is due to how interactions were happening with the
xforms to glue async between them.

*Architectural Model did not fit*

The xforms implementation assumed that xforms had no knowledge of the
system they were executing up until the point of reduction.  This caused
the use of non-standard rfs.  In order for async to function, it must
operate on a stack of functions which become glued together with some
form of continuation.

*Unable to to stop & start in single pass*

This was an interesting observation, execution under xforms passed over
components in a particular order.  This made it harder to operate like:

[start A] [start B] [start C] [stop A] [stop B] [stop C]

Instead, only interposed operations were supported:

[start A] [stop A] [start B] [stop B] [start C] [stop C]