Rewrite clojure interpreter
The old interpreter did not correctly handle maps, and had ambiguous
rules around resolving symbols.  Now symbols are only resolved in lists,
making the use of symbols as keys much easier.  It also ensures that
metadata is preserved on vectors and maps.

There are now fairly comprehensive tests for the kinds of thing I expect
to work.