Set value from :stop process
This cleans up two keys that were leftover experimentation (is it
possible to set two keys at the same time, even for async?).  It also
starts setting the :stop to the return value of the :stop function, this
is a testing aid for now, but might also be useful for real programs.