Fix nested evaluations inside non-seq structures
This was caused by `(list identity {:x (list inc 1)})`.  The
`(list inc 1)` resulted in a prevent-eval value, that would never be
eval'd.  This is because on the way out, the map was get-value'd, but
it's child values were not.  This meant that the `(identity)` call would
get a map with a prevent-eval'd function in.