Add alpha reloadable functionality
Without clip, and with req injection, you're generally able to setup so
that no action is required after re-evaluating a function.  This is
mostly a dev-time convenience and makes it much easier to do
development.

Now, a :reloads is provided which can be used to specify a type or
component that a reload configuration exists for.